Big chunk of work on API
This change adds the encoder and decoder to Settings. That means when
you deal with a cache, it must hold a certain type of value. So it is
(Cache Model) or (Cache Student) so you are more likely to keep things
organized.

This change also contains draft implementations for all public
functions. There are definitely some problems, but it is moving towards
compiling.